+++ DOM/components/checkComplete.js
@@ -1,0 +1,17 @@
+const checkComplete = () => {
+  const i = document.createElement("i");
+  i.classList.add("far", "fa-check-square", "icon");
+  i.addEventListener("click", completeTask);
+  return i;
+};
+
+// IIFE
+const completeTask = (event) => {
+  // target and toggle
+  const element = event.target;
+  element.classList.toggle("fas");
+  element.classList.toggle("completeIcon");
+  element.classList.toggle("far");
+};
+
+export default checkComplete;
